chore: Move auth endpoints into separate modules
