+++ server/db.ts
@@ -9,7 +9,9 @@ import { array, enums, nullable, object, optional, union } from "superstruct";
 import { BadRequestError, NotEnoughRoomError, NotFoundError } from "./errors/index.js";
 import { destroyFileData, fetchFileData, statsForUser, upsertFileData } from "./database/io.js";
 import { handleErrors } from "./handleErrors.js";
-import { maxSpacePerUser, MAX_FILE_BYTES, ownersOnly, requireAuth } from "./auth/index.js";
+import { maxSpacePerUser, MAX_FILE_BYTES } from "./auth/limits.js";
+import { ownersOnly } from "./auth/ownersOnly.js";
+import { requireAuth } from "./auth/requireAuth.js";
 import { respondData, respondSuccess } from "./responses.js";
 import { sep as pathSeparator } from "node:path";
 import { Router } from "express";
